refactor: cleaner separation of concerns in player classes
